+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## UPCOMING
 
 * Update `VerifyUserEmailView` to redirect to login without providing a next.
+* Redirect to the login when attempting to verify an email address that is already verified.
 
 ## 15.0.0